+++ src/main/scala/KamonJfr.scala
@@ -1,17 +1,17 @@
 import com.typesafe.config.Config
 import jdk.jfr.Configuration
-import jdk.jfr.consumer.{EventStream, RecordedEvent, RecordingStream, MetadataEvent}
+import jdk.jfr.consumer.{EventStream, MetadataEvent, RecordedEvent, RecordingStream}
 import kamon.Kamon
 import kamon.metric.PeriodSnapshot
 import kamon.module.MetricReporter
+import metrics.{Cpu, GarbageCollection, Memory, Threads}
 
 import java.time.Duration
 import java.util.function.Consumer
 import scala.collection.concurrent.TrieMap
 import scala.util.Using
 import java.time.Instant
-import scala.jdk.CollectionConverters._
-
+import scala.jdk.CollectionConverters.*
 
 object KamonJfr {
   @volatile var jfrConfig: Map[String, String] = jfrConfigFrom(Kamon.config())
@@ -52,14 +52,14 @@ object KamonJfr {
 
     def onEvent(event: RecordedEvent): Unit = {
       event.getEventType.getName match {
-        case "jdk.CPULoad" => CpuHandler.onCPULoad(event)
-        case "jdk.PhysicalMemory" => MemoryHandler.onPhysicalMemory(event)
-        case "jdk.GarbageCollection" => GCHandler.onGarbageCollection(event)
+        case "jdk.CPULoad" => Cpu.onCPULoad(event)
+        case "jdk.PhysicalMemory" => Memory.onPhysicalMemory(event)
+        case "jdk.GarbageCollection" => GarbageCollection.onGarbageCollection(event)
+        case "jdk.JavaThreadStatistics" => Threads.onJavaThreadStatistics(event)
 
         // case "jdk.JavaMonitorEnter" => JavaMonitorHandler.onMonitorEnter(event)
-        // case "jdk.JavaThreadStatistics" => Threads.onJavaThreadStatistics(event)
-        // case "jdk.SafepointBegin" => SafePointHandler.onSafepointBegin(event)
-        // case "jdk.SafepointEnd" => SafePointHandler.onSafepointEnd(event)
+//         case "jdk.SafepointBegin" => SafePointHandler.onSafepointBegin(event)
+//         case "jdk.SafepointEnd" => SafePointHandler.onSafepointEnd(event)
         // case "jdk.ExecutionSample" => println("====>>>> ExecutionSample"+ event)
         // case "jdk.ObjectAllocationSample" => println("====>>>> allocationSample"+ event)
         // case "jdk.NetworkUtilization" => println("====>>>> jdk.NetworkUtilization"+ event)
@@ -73,34 +73,4 @@ object KamonJfr {
     Thread.sleep(10000000)
   }
 }
-
-
- 
-
-
-
-// val safepointBegin = TrieMap.empty[Long, Instant]
-
-//  
-//  System.gc()
-
-
-
 ////https://github.com/DataDog/dd-trace-java/blob/master/dd-java-agent/agent-profiling/profiling-controller-openjdk/src/main/java/com/datadog/profiling/controller/openjdk/OpenJdkController.java
-//
-
- //  es.enable("jdk.CPULoad").withPeriod(duration)
-  //  es.enable("jdk.YoungGarbageCollection").withoutThreshold
-  //  es.enable("jdk.OldGarbageCollection").withoutThreshold
-  //  es.enable("jdk.GCHeapSummary").withPeriod(duration)
-  //  es.enable("jdk.PhysicalMemory").withPeriod(duration)
-  //  es.enable("jdk.GCConfiguration").withPeriod(duration)
-  //  es.enable("jdk.SafepointBegin")
-  //  es.enable("jdk.SafepointEnd")
-  //  es.enable("jdk.ObjectAllocationSample").`with`("throttle", "150/s")
-  //  es.enable("jdk.ExecutionSample").withPeriod(Duration.ofMillis(10)).withStackTrace
-  //  es.enable("jdk.JavaThreadStatistics").withPeriod(duration)
-  //  es.enable("jdk.ClassLoadingStatistics").withPeriod(duration)
-  //  es.enable("jdk.Compilation").withoutThreshold
-  //  es.enable("jdk.GCHeapConfiguration").withPeriod(duration)
-  //  es.enable("jdk.Flush").withoutThreshold
